Remove idle topic finder
